<==== Mercurial data ====>
Branch: ft_208_wearable_camera_upload_test

+++ libs/nx_sql/src/nx/sql/sql_query_execution_helper.cpp
@@ -28,7 +28,7 @@ QList<QByteArray> quotedSplit(const QByteArray& data)
         else if (*curPtr == ';' && !inQuote1 && !inQuote2)
         {
             //*curPtr = 0;
-            result << QByteArray::fromRawData(prevPtr, curPtr - prevPtr);
+            result << QByteArray::fromRawData(prevPtr, (int) (curPtr - prevPtr));
             prevPtr = curPtr + 1;
         }
     }
